==> nbs/01_model_building/gtex/00_gtex.ipynb <==
# GTEx model building with CLAMP and PLIER

💡 **Environment:** `clamp-analyses`  

This notebook builds latent variable models from GTEx v8 RNA‑seq TPM data using CLAMP and PLIER. It automates downloading and preprocessing the GTEx matrix, creates a Filebacked Big Matrix (FBM), computes an SVD to estimate the model dimension, prepares pathway priors, runs CLAMP (base + full) and PLIER, and saves model outputs (B, Z, summaries) and intermediate files. Configuration and paths are controlled via `config.R`.

# CLAMP

## Load libraries

# Create a timestamp to track the start of the analysis
start_time <- Sys.time()
cat("GTEx CLAMP and PLIER analysis started at:", format(start_time), "\n")

if (!requireNamespace("PLIER", quietly = TRUE)) {
    devtools::install_github("wgmao/PLIER")
}

library(bigstatsr)
library(data.table)
library(dplyr)
library(rsvd)
library(glmnet)
library(Matrix)
library(knitr)
library(here)
library(PLIER)
library(CLAMP)

source(here("config.R"))

set.seed(config$GTEx$RANDOM_SVD_SEED)

## Output directory

output_data_dir <- config$GTEx$DATASET_FOLDER
dir.create(output_data_dir, showWarnings = FALSE, recursive = TRUE)

# Settings

block_size <- config$GENERAL$CHUNK_SIZE
N_CORES    <- config$GTEx$N_CORES

## Download GTEx 

url <- config$GTEx$URL
dest_dir <-  config$GTEx$DATASET_FOLDER
dest_gz  <- file.path(dest_dir, basename(url))

if (!file.exists(dest_gz)) {
  dir.create(dest_dir, recursive = TRUE, showWarnings = FALSE)
  download.file(url, dest_gz, mode = "wb")
  message("Downloaded to: ", dest_gz)
} else {
  message("File already exists, skipping download.")
}

## Preprocess GTEx data

exprs_path  <- file.path(config$GTEx$DATASET_FOLDER, 'GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_tpm.gct.gz')
output_file <- config$GTEx$DATASET_FILE

if (!file.exists(output_file)) {
  dir.create(dirname(output_file), recursive = TRUE, showWarnings = FALSE)
  exprs_data <- read.table(exprs_path, header = TRUE, sep = "\t", skip = 2, check.names = FALSE)
  saveRDS(exprs_data, config$GTEx$DATASET_FILE)
  message("File successfully written to: ", config$GTEx$DATASET_FILE)
} else {
  message("Output file already exists. Skipping.")
}

# Aggregate in-place by 'description'
gtex <- readRDS(here(config$GTEx$DATASET_FILE))
gtex <- as.data.table(gtex)
aggregated_gtex <- gtex[, lapply(.SD, sum), by = Description, .SDcols = is.numeric]

genes <- aggregated_gtex$Description
samples <- colnames(aggregated_gtex[, -1])
data_mat <- as.matrix(aggregated_gtex[, -1])

fbm_files <- c(
  file.path(output_data_dir, "FBMgtex.bk"),
  file.path(output_data_dir, "FBMgtex.bk_preproc.bk"),
  file.path(output_data_dir, "FBMgtex.bk_preproc_filtered.bk"),
  file.path(output_data_dir, "FBMgtex.bk_preproc_filtered.rds")
)

for (f in fbm_files) {
  if (file.exists(f)) {
    res <- unlink(f)
    if (identical(res, 0L)) {
      message("✅ Removed existing FBM file: ", f)
    } else {
      warning("⚠️ Failed to remove: ", f)
    }
  } else {
    message("ℹ️ Not found, skipping: ", f)
  }
}

# Create the FBM
fbm_file <- file.path(output_data_dir, "FBMgtex")
gtexFBM <- FBM(nrow = nrow(data_mat), ncol = ncol(data_mat), backingfile = fbm_file, create_bk = T)

# Populate it with data
n_blocks <- ceiling(nrow(aggregated_gtex) / block_size)

for (i in 1:n_blocks) {
  start_row <- (i-1) * block_size + 1
  end_row <- min(i * block_size, nrow(data_mat))
  
  gtexFBM[start_row:end_row, ] <- as.matrix(data_mat[start_row:end_row, ])
}

# Preprocess and z‑score FBM

prep_gtex <- preprocessCLAMPFBM(
  fbm        = gtexFBM,
  mean_cutoff= config$GTEx$GENES_MEAN_CUTOFF,
  var_cutoff = config$GTEx$GENES_VAR_CUTOFF,
  ncores=N_CORES
)

gtex_fbm_filt <- prep_gtex$fbm_filtered
gtex_rowStats <- prep_gtex$rowStats

Check overlap genes with GTEx v8 RNA-seq data and TWAS gene list

twas_genes <- read.csv(here('output/creating_twas_gwas_list/gene_list_union_mashr_and_elastic_net_with_phi.csv'))
head(twas_genes)

overlap_genes <- intersect(genes[prep_gtex$kept_rows], twas_genes$Gene_Symbol)

percent_overlap <- (length(overlap_genes) / length(twas_genes$Gene_Symbol)) * 100

cat("Number of overlapping genes:", length(overlap_genes), "\n")
cat("Percentage overlap (relative to TWAS genes):", round(percent_overlap, 2), "%\n")

zscoreCLAMPFBM(gtex_fbm_filt, gtex_rowStats, ncores=N_CORES)
gtex_genes <- genes[prep_gtex$kept_rows]

saveRDS(samples, file = file.path(output_data_dir, "gtex_samples.rds"))

saveRDS(gtex_genes, file = file.path(output_data_dir, "gtex_genes.rds"))

saveRDS(gtex_fbm_filt, file = file.path(output_data_dir, "gtex_fbm_filt.rds"))

df_gtex_fbm_filt <- as.data.frame(as.matrix(gtex_fbm_filt[]))

colnames(df_gtex_fbm_filt) <- samples

rownames(df_gtex_fbm_filt) <- gtex_genes

head(df_gtex_fbm_filt)

saveRDS(df_gtex_fbm_filt, file = file.path(output_data_dir, "df_gtex_fbm_filt.rds"))
write.csv(df_gtex_fbm_filt, file = file.path(output_data_dir, "df_gtex_fbm_filt.csv"))

## SVD computation

if (!file.exists(file.path(output_data_dir, "gtex_svdRes.rds"))) {

  g_fb <- nrow(gtex_fbm_filt)
  samples_fb <- ncol(gtex_fbm_filt)
  SVD_K_gtex <- min(g_fb, samples_fb) - 1

  message("Using SVD K = ", SVD_K_gtex)

  if (N_CORES > 1) {
    # disable BLAS parallelization when parallelizing
    options(bigstatsr.check.parallel.blas = FALSE)
    blas_nproc <- getOption("default.nproc.blas")
    options(default.nproc.blas = NULL)
  }

  gtex_svdRes <- big_randomSVD(
    gtex_fbm_filt,
    k = SVD_K_gtex,
    ncores = N_CORES
  )

  saveRDS(gtex_svdRes, file = file.path(output_data_dir, "gtex_svdRes.rds"))

  if (N_CORES > 1) {
    # restore previous state
    options(bigstatsr.check.parallel.blas = TRUE)
    options(default.nproc.blas = blas_nproc)
  }

} else {
  message("gtex_svdRes already exists, skipping SVD computation.")
}

## Estimate K for CLAMP

gtex_fbm_filt <- readRDS(file.path(output_data_dir, "gtex_fbm_filt.rds"))
gtex_svdRes <- readRDS(file.path(output_data_dir, "gtex_svdRes.rds"))

CLAMP_K_gtex <- num.pc(list(d = gtex_svdRes$d)) * 2
message("Inferred CLAMP K = ", CLAMP_K_gtex)

saveRDS(CLAMP_K_gtex, file = file.path(output_data_dir, "CLAMP_K_gtex.rds"))

## CLAMPbase

gtex_baseRes <- CLAMPbase(
  Y      = gtex_fbm_filt,
  svdres = gtex_svdRes,
  trace  = TRUE,
  clamp_k = CLAMP_K_gtex
)

gtex_baseRes$Z <- data.frame(gtex_baseRes$Z)
rownames(gtex_baseRes$Z) <- gtex_genes
head(gtex_baseRes$Z)

gtex_baseRes$B <- data.frame(gtex_baseRes$B)
colnames(gtex_baseRes$B) <- samples
head(gtex_baseRes$B)

saveRDS(gtex_baseRes, file = file.path(output_data_dir, "CLAMPbase.rds"))

model_dir <- file.path(output_data_dir, "CLAMPbase")
dir.create(model_dir, showWarnings = FALSE, recursive = TRUE)

B <- gtex_baseRes$B
write.csv(B, file.path(model_dir, "B.csv"))

Z <- gtex_baseRes$Z
write.csv(Z, file.path(model_dir, "Z.csv"))

## Prepare pathway priors

gtex_gmtList <- list(
  KEGG = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=KEGG_2021_Human"),
  BP = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=GO_Biological_Process_2025"),
  GTEx_Tissues = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=GTEx_Tissues_V8_2023")
)

# prefix each gene‐set name with its library to guarantee uniqueness
for(lib in names(gtex_gmtList)) {
  names(gtex_gmtList[[lib]]) <- paste0(lib, "_", names(gtex_gmtList[[lib]]))
}

gtex_pathMat <- gmtListToSparseMat(gtex_gmtList)
gtex_matched <- getMatchedPathwayMat(gtex_pathMat, gtex_genes)
gtex_chatObj <- getChat(gtex_matched)

## CLAMPfull

gtex_fullRes <- CLAMPfull(
    Y = gtex_fbm_filt,
    svdres = gtex_svdRes,
    priorMat = gtex_matched,
    clamp.base.result = gtex_baseRes,
    use_cpp = TRUE,
    trace = TRUE,
    clamp_k = CLAMP_K_gtex
  )

gtex_baseRes$Z <- data.frame(gtex_baseRes$Z)
rownames(gtex_baseRes$Z) <- gtex_genes
head(gtex_baseRes$Z)

gtex_baseRes$B <- data.frame(gtex_baseRes$B)
colnames(gtex_baseRes$B) <- samples
head(gtex_baseRes$B)

gtex_fullRes$summary <- gtex_fullRes$summary %>%
    dplyr::rename(LV = LV_index)  %>% 
    dplyr::mutate(LV = paste0('LV', LV))

saveRDS(gtex_fullRes, file = file.path(output_data_dir, "CLAMPfull.rds"))

model_dir <- file.path(output_data_dir, "CLAMPfull")
dir.create(model_dir, showWarnings = FALSE, recursive = TRUE)

B <- gtex_fullRes$B
write.csv(B, file.path(model_dir, "B.csv"))

Z <- gtex_fullRes$Z
write.csv(Z, file.path(model_dir, "Z.csv"))

summary <- gtex_fullRes$summary
write.csv(summary, file.path(model_dir, "summary.csv"))

# Create a timestamp to track the end of the analysis
end_time <- Sys.time()
elapsed_time <- end_time - start_time
cat("GTEx CLAMP analysis completed at:", format(end_time), "\n")
cat("Total elapsed time:", format(elapsed_time), "\n")

# PLIER

Run PLIER with the same inputs

gtex_plier = PLIER::PLIER(
    gtex_fbm_filt[], 
    as.matrix(gtex_matched), 
    svdres = gtex_svdRes, 
    Chat = as.matrix(gtex_chatObj), 
    doCrossval = TRUE, 
    k = CLAMP_K_gtex
  )

colnames(gtex_plier$Z) <- paste0('LV', seq_len(ncol(gtex_plier$Z)))

head(gtex_plier$Z)
dim(gtex_plier$Z)

gtex_plier$summary <- gtex_plier$summary %>%
    dplyr::rename(LV = `LV index`)  %>% 
    dplyr::mutate(LV = paste0('LV', LV))

head(gtex_plier$summary)

head(gtex_plier$B)
dim(gtex_plier$B)

colnames(gtex_plier$B) <- samples

saveRDS(gtex_plier, file = file.path(output_data_dir, "PLIER.rds"))

model_dir <- file.path(output_data_dir, "PLIER")
dir.create(model_dir, showWarnings = FALSE, recursive = TRUE)

B <- gtex_plier$B
write.csv(B, file.path(model_dir, "B.csv"))

Z <- gtex_plier$Z
write.csv(Z, file.path(model_dir, "Z.csv"))

summary <- gtex_plier$summary
write.csv(summary, file.path(model_dir, "summary.csv"))
