==> nbs/01_model_building/gtex/05_gtex_nmf_pca.ipynb <==
# GTEx model building with NFM

## Libraries

import csv
import gc
from pathlib import Path
import numpy as np
import pandas as pd
from pyprojroot.here import here
from sklearn.cluster import KMeans
from sklearn.metrics import adjusted_rand_score as ari
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import adjusted_rand_score
import cupy as cp
import numpy as np
import pandas as pd
from sklearn.decomposition import PCA
from sklearn.decomposition import NMF
from sklearn.decomposition import FastICA

## Input

gtex_data = pd.read_csv(here('output/gtex/df_gtex_fbm_filt.csv'), index_col=0)

gtex_data.head()

n_components = 412

# PCA

pca = PCA(n_components=n_components, svd_solver="auto", random_state=0)

W = pca.fit_transform(gtex_data.T)     
H = pca.components_         

pc_names = [f"PC{i+1}" for i in range(W.shape[1])]

gtex_pca_scores = pd.DataFrame(W, index=gtex_data.columns, columns=pc_names)
gtex_pca_loadings = pd.DataFrame(H, index=pc_names, columns=gtex_data.index)

gtex_pca_B = gtex_pca_scores.T 
gtex_pca_B.index.name = "PC"

gtex_pca_B.head()

gtex_pca_B.to_pickle(here('output/gtex/gtex_pca_B.pkl'))

# ICA

ica = FastICA(n_components=n_components, random_state=0, max_iter=1000)

W = ica.fit_transform(gtex_data.T)      
H = ica.mixing_.T                      

ic_names = [f"IC{i+1}" for i in range(W.shape[1])]

gtex_ica_scores = pd.DataFrame(W, index=gtex_data.columns, columns=ic_names)
gtex_ica_loadings = pd.DataFrame(H, index=ic_names, columns=gtex_data.index)

gtex_ica_B = gtex_ica_scores.T
gtex_ica_B.index.name = "IC"

gtex_ica_B.head()

gtex_ica_B.to_pickle(here('output/gtex/gtex_ica_B.pkl'))

# NMF  

nmf = NMF(
    n_components=n_components,
    init="nndsvd",
    random_state=0,
    max_iter=500,
)

W = nmf.fit_transform(gtex_data.T)
H = nmf.components_

comp_names = [f"LV{i+1}" for i in range(W.shape[1])]

gtex_nmf_scores = pd.DataFrame(W, index=gtex_data.index, columns=comp_names)
gtex_nmf_loadings = pd.DataFrame(H, index=comp_names, columns=gtex_data.index)
gtex_nmf_B = gtex_nmf_scores.T
gtex_nmf_B.index.name = "LV"

gtex_nmf_B.to_pickle(here('output/gtex/gtex_nmf_B.pkl'))
