==> nbs/01_model_building/gtex/07_flashier.ipynb <==
# GTEx model building with flashier

Empirical Bayes non-negative matrix factorization for single-cell RNA-seq data

## Libraries

if (!requireNamespace("flashier", quietly = TRUE)) {
  install.packages(
  "fastTopics",
  repos = c("https://stephenslab.r-universe.dev", "https://cloud.r-project.org")
  )
  remotes::install_github("willwerscheid/flashier", dependencies = c("Depends","Imports"))
}

library(flashier)
library(here)
library(flashier)
library(Matrix)
library(fastTopics)
library(ggplot2)
set.seed(3)

## Input

gtex_data <- readRDS(here('output/gtex/df_gtex_fbm_filt.rds'))
head(gtex_data)

K <- 412

# Run Flashier

flash_gtex <- flash(gtex_data,
    ebnm_fn = ebnm_normal,
    backfit = TRUE,
    greedy_Kmax = K      
)

plot(flash_gtex, pm_which = "factors", pm_colors = gtex_colors, plot_type = "bar")
