==> nbs/01_model_building/gtex/08_mofa.ipynb <==
# GTEx model building with MOFA2

Empirical Bayes non-negative matrix factorization for single-cell RNA-seq data

## Libraries

library(data.table)
library(here)
library(MOFA2)
library(dplyr)

# Input

gtex_data <- readRDS(here('output/gtex/df_gtex_fbm_filt.rds'))
head(gtex_data)

# Create model

X <- as.matrix(gtex_data)

data_list <- list(RNA = X)
mofa <- create_mofa(data_list)

data_opts  <- get_default_data_options(mofa)
# your data are already normalized; don't rescale views/groups
data_opts$scale_views  <- FALSE
data_opts$scale_groups <- FALSE

model_opts <- get_default_model_options(mofa)
model_opts$num_factors <- 412  

train_opts <- get_default_training_options(mofa)
train_opts$convergence_mode <- "fast" # quick run
train_opts$maxiter <- 1000            # default; you can lower to ~500 for speed

set.seed(1)
MOFAobject <- prepare_mofa(
  object = mofa,
  data_options = data_opts,
  model_options = model_opts,
  training_options = train_opts
)

MOFAobject

outfile = file.path(tempdir(),"model.hdf5")
MOFAobject.trained <- run_mofa(MOFAobject, outfile, use_basilisk=TRUE)
