==> nbs/01_model_building/gtex/09_cogaps.ipynb <==
# CoGAPs

library(CoGAPS)
library(here)

## Input

gtex_data <- readRDS(here('output/gtex/df_gtex_fbm_filt.rds'))
head(gtex_data)

gtex_data_shifted <- gtex_data - min(gtex_data)

K <- 412

## CoGAPs run

params <- new("CogapsParams", nPatterns=K)

cogapsresult <- CoGAPS(gtex_data_shifted, params, outputFrequency = 10000)
